requirements.txt
aiogram~=2.22.2
librosa~=0.9.2
google-cloud-storage~=2.5.0
google-cloud-speech~=2.16.2